assistant_example/requirements.txt
fastapi==0.104.1
matplotlib==3.7.2
openai==1.3.5
Pillow==9.3.0
Pillow==10.1.0
pydantic==1.10.12
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.24.0.post1

# pip3 install -r requirements.txt
